requirements.txt
aiogram
python-dotenv
telethon